requirements.txt
streamlit
requests
pydub
transformers
torch
fpdf
python-dotenv
